add material design symbols
closes #3.

the material-symbols submodule is set to point to commit
ea15af63efe0d712a7312a3a47d1a31d031d0dc5. later commits changed the
SVG format in a way that is seemingly incompatible with
GTK -- using the latest revision will result in the icons being
included, but not rendering.